requirements.txt
anyio==3.6.2
asyncio==3.4.3
beautifulsoup4==4.12.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confluent-kafka==2.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
filelock==3.10.7
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-logging==3.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
lxml==4.9.2
MarkupSafe==2.1.2
orjson==3.8.11
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
PySocks==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.2
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@285d5874fc20a9d9463ed261f45c5f4118277d05
soupsieve==2.4
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
